move ReportingPeriodsCell files into ReportingPeriods folder, remove duplicate ReportingPeriodsCell files, add mock data for storybook
