+++ web/src/components/ReportingPeriod/ReportingPeriodsCell/ReportingPeriodsCell.tsx
@@ -1,44 +1,48 @@
-import type { FindReportingPeriods } from 'types/graphql'
+import type { FindReportingPeriodsWithCertification } from 'types/graphql'
 
-import { Link, routes } from '@redwoodjs/router'
 import type { CellSuccessProps, CellFailureProps } from '@redwoodjs/web'
 
 import ReportingPeriods from 'src/components/ReportingPeriod/ReportingPeriods'
 
 export const QUERY = gql`
-  query FindReportingPeriods {
-    reportingPeriods {
+  query FindReportingPeriodsWithCertification {
+    reportingPeriodsWithCertification {
       id
       name
       startDate
       endDate
-      inputTemplateId
-      outputTemplateId
-      createdAt
-      updatedAt
+      certificationForOrganization {
+        id
+        createdAt
+        certifiedBy {
+          id
+          email
+        }
+      }
+    }
+    organizationOfCurrentUser {
+      id
+      preferences
     }
   }
 `
 
 export const Loading = () => <div>Loading...</div>
 
-export const Empty = () => {
-  return (
-    <div className="rw-text-center">
-      {'No reportingPeriods yet. '}
-      <Link to={routes.newReportingPeriod()} className="rw-link">
-        {'Create one?'}
-      </Link>
-    </div>
-  )
-}
+export const Empty = () => <div>Empty</div>
 
 export const Failure = ({ error }: CellFailureProps) => (
-  <div className="rw-cell-error">{error?.message}</div>
+  <div style={{ color: 'red' }}>Error: {error?.message}</div>
 )
 
 export const Success = ({
-  reportingPeriods,
-}: CellSuccessProps<FindReportingPeriods>) => {
-  return <ReportingPeriods reportingPeriods={reportingPeriods} />
+  reportingPeriodsWithCertification,
+  organizationOfCurrentUser,
+}: CellSuccessProps<FindReportingPeriodsWithCertification>) => {
+  return (
+    <ReportingPeriods
+      reportingPeriods={reportingPeriodsWithCertification}
+      organizationOfCurrentUser={organizationOfCurrentUser}
+    />
+  )
 }
