--- web/src/components/ReportingPeriodsCell/ReportingPeriodsCell.mock.ts
@@ -1,4 +1,0 @@
-// Define your own mock data here:
-export const standard = (/* vars, { ctx, req } */) => ({
-  reportingPeriods: [{ id: 42 }, { id: 43 }, { id: 44 }],
-})
